Rename Request & Query factory class
