+++ Doctrine/Orm/Extension/PaginationExtension.php
@@ -17,7 +17,7 @@ use Doctrine\ORM\Tools\Pagination\Paginator as DoctrineOrmPaginator;
 use Dunglas\ApiBundle\Api\ResourceInterface;
 use Dunglas\ApiBundle\Doctrine\Orm\Paginator;
 use Dunglas\ApiBundle\Doctrine\Orm\QueryResultExtensionInterface;
-use Dunglas\ApiBundle\Doctrine\Orm\Util\QueryFactory;
+use Dunglas\ApiBundle\Doctrine\Orm\Util\QueryChecker;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
@@ -155,7 +155,7 @@ class PaginationExtension implements QueryResultExtensionInterface
          *
          * @see https://github.com/doctrine/doctrine2/blob/900b55d16afdcdeb5100d435a7166d3a425b9873/lib/Doctrine/ORM/Tools/Pagination/CountWalker.php#L50
          */
-        if (QueryFactory::hasHavingClause($queryBuilder)) {
+        if (QueryChecker::hasHavingClause($queryBuilder)) {
             return true;
         }
 
@@ -164,7 +164,7 @@ class PaginationExtension implements QueryResultExtensionInterface
          *
          * @see https://github.com/doctrine/doctrine2/blob/900b55d16afdcdeb5100d435a7166d3a425b9873/lib/Doctrine/ORM/Tools/Pagination/LimitSubqueryWalker.php#L87
          */
-        if (QueryFactory::hasRootEntityWithForeignKeyIdentifier($queryBuilder, $this->managerRegistry)) {
+        if (QueryChecker::hasRootEntityWithForeignKeyIdentifier($queryBuilder, $this->managerRegistry)) {
             return true;
         }
 
@@ -174,8 +174,8 @@ class PaginationExtension implements QueryResultExtensionInterface
          * @see https://github.com/doctrine/doctrine2/blob/900b55d16afdcdeb5100d435a7166d3a425b9873/lib/Doctrine/ORM/Tools/Pagination/LimitSubqueryWalker.php#L149
          */
         if (
-            QueryFactory::hasMaxResults($queryBuilder)
-            && QueryFactory::hasOrderByOnToManyJoin($queryBuilder, $this->managerRegistry)
+            QueryChecker::hasMaxResults($queryBuilder)
+            && QueryChecker::hasOrderByOnToManyJoin($queryBuilder, $this->managerRegistry)
         ) {
             return true;
         }
